test: add radix cross-browser e2e coverage
